Update const.py
During the development of a tool, we found out that the hooks for CoReleaseMarshalData and ImageListAdd where not working. Apparently, these functions are properly declared in Qiling Framework version 1.2, but when we try to run our tool we got a “Windows API implementation error“. This is because Qiling is missing some references in its const.py file. More precisely, it is missing the references for:

"HBITMAP": "HANDLE",
"HIMAGELIST": "HANDLE",
"LPSTREAM" : "POINTER"