Start implementation of sys_conf message handling
The first message type of the System Manager is sys_conf. This code handles
the sys_conf(GET, STRING, LIBPATH) case. This is needed to run dynamically
linked programs because the loader issuses this call to get the folder where
he searches for the required libraries.
The path /usr/lib of the rootfs is hard-coded here.