Do not double align the length
If we do a mmap(len = 4096, prot = PROT_NONE), get a memory page
from 4096 to 8191, and then call mmap(addr = 5120, len = 1024,
prot = PROT_READ) to change the page protection, then the effective
size 4096 is calculated and used for the call to protect with
ql.mem.protect(addr = 5120, len = 4096, prot = PROT_READ).
In mem.protect the address is aligned to 4096 and the aligned size
is used to calculate the new aligned size of 5120. This results in
an exception because the final uc.mem_protect() call excepts the
size of the page that was mapped with the first mmap() call.
Because ql.mem.protect() performs the alignment we can use the real
size instead of the effective mmap size.