Only zero-initialize memory pages once
If mmap is called to obtain a memory region and later called for
the same memory region to change the page protection then the
current implementation overwrites the whole region with zeros.
This commit changes the behavior to only zero at newly allocated
pages and not pages that were already used by the program.